# contracts/openapi/phase-01.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "browserSession": {
        "in": "cookie",
        "name": "__Host-hdi-session",
        "type": "apiKey"
      },
      "serviceBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hospital Data Intelligence Governance API",
    "version": "0.0.0-phase-01"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/callback": {
      "get": {
        "operationId": "completeBrowserLogin",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 32,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "iss",
            "required": false,
            "schema": {
              "maxLength": 2048,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_state",
            "required": false,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [],
        "summary": "完成服务端OIDC代码交换并建立不透明会话"
      }
    },
    "/auth/login": {
      "get": {
        "operationId": "beginBrowserLogin",
        "parameters": [
          {
            "in": "query",
            "name": "returnTo",
            "required": false,
            "schema": {
              "pattern": "^/(?!/).*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [],
        "summary": "启动服务端Authorization Code与PKCE S256登录"
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logoutBrowserSession",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "minLength": 32,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "loggedOut": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "loggedOut"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ],
        "summary": "撤销当前服务端会话"
      }
    },
    "/auth/session": {
      "get": {
        "operationId": "getBrowserSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "csrfToken": {
                      "minLength": 32,
                      "type": "string"
                    },
                    "principalId": {
                      "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                      "type": "string"
                    },
                    "principalKind": {
                      "enum": [
                        "PERSON"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "principalId",
                    "principalKind",
                    "csrfToken"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ],
        "summary": "读取当前服务端会话和CSRF请求令牌"
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}